fix: create-fabrice-ai fixed to use *.workflow.ts and *.test.ts files for the examples
